+++ CHANGELOG.md
@@ -30,6 +30,7 @@ See [STATUS.md](server/STATUS.md) to learn more about which features will remain
 - [#851](https://github.com/ontola/atomic-server/issues/851) Deleting file resources now also deletes the file from the filesystem.
   BREAKING: [#1107](https://github.com/ontola/atomic-server/issues/1107) Named nested resources are no longer supported. Value::Resource and SubResource::Resource have been removed. If you need to include multiple resources in a response use an array.
   BREAKING: `store.get_resource_extended()` now returns a `ResourceResponse` instead of a `Resource` due to the removal of named nested resources. Use `.into()` or `.to_single()` to convert to a `Resource`.
+- [#415](https://github.com/ontola/atomic-server/issues/415) Mutli-filter queries.
 
 ## [v0.40.2]